Changed the version data type

+++ protocols/ofp10/protocol.go
@@ -8,7 +8,7 @@ import (
 
 const (
 	// Version is the value of version byte in ofp header
-	Version = 0x01
+	Version = uint8(0x01)
 )
 
 // OfpConfigFlags
