+++ protocols/ofp11/protocol.go
@@ -2,5 +2,5 @@ package ofp11
 
 const (
 	// Version is the value of version byte in ofp header
-	Version = 0x02
+	Version = uint8(0x02)
 )
